Create the base class for placing the common logic.
